Issue fix  Telephony duration units #778 Duration to be reported as milli seconds https://github.com/BIDMCDigitalPsychiatry/LAMP-platform/issues/778#top Gradle updates

+++ lamp_kotlin/src/main/java/digital/lamp/lamp_kotlin/sensor_core/TelephonySensor.kt
@@ -28,8 +28,8 @@ class TelephonySensor : Service() {
       //  fun onIncomingCallReceived(ctx: Context?, start: Long?)
        // fun onIncomingCallAnswered(ctx: Context?, start: Long?)
       // fun onOutgoingCallStarted( start: Long?)
-        fun onIncomingCallEnded(callDuration:Int?)
-        fun onOutgoingCallEnded( callDuration:Int?)
+        fun onIncomingCallEnded(callDuration:Long?)
+        fun onOutgoingCallEnded( callDuration:Long?)
         fun onMissedCall()
     }
